requirements.txt
Flask==2.3.2
Flask-Mail==0.9.1
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Flask-WTF==1.0.1
python-dotenv==1.0.0
Flask-CSRFProtect==1.1.0
requests==2.28.1
Werkzeug==2.3.0
SQLAlchemy==2.0.0
Flask-SQLAlchemy==3.0.2
gunicorn==21.2.0